Modify 'ROW_LENGTH' and 'COLUMN_LENGTH' variables
Since the size of the grid will change with the user's input, the
'ROW_LENGTH' and 'COLUMN_LENGTH' variables will need to change from
being defined as 'const' to 'let'. Consequently, their naming schemes
will no longer be suitable, thus the names for the variables
representing the row and column length will be renamed.