Refactor the loop generating the grid
With the user being able to change the size of the grid based on their
input, the loop generating the initial grid will be performing the same
task when generating a new grid, just with different size values. Thus,
the loop generating the grid was refactored into its own function.