Refactor grid array and grid-related functions
With how the grid array is currently structured and stored in JS, the
CSS is not able to evenly distribute the given size of squares on both
sides. In order to achieve this, another set of div elements that act as
the 'columns' of the grid need to be added, so that the container will
have 'x' amount of columns, with each column having 'x' amount of
'rows', where 'x' is the given size of the grid.

As a result, the way div elements were stored when forming the grid had
to be changed to accommodate for both 'column' and 'row' divs, which
resulted in using two arrays instead of one. Alongside this, functions
related to manipulating the grid (i.e., dissolving and creating a grid)
were also refactored to accommodate for this change.