Add styles to div elements related in grid making
To align with the changes made in the script, changes were also made to
the stylesheet (changing name of class selector, adding new class
selector).

Furthermore, relevant flexbox properties were added to each div element
so that when the grid forms, the grid will always occupy the same total
space on the screen, but the size of the squares will adapt to that same
total space depending on the number of squares per side.